\name{medianVariogram}
\alias{medianVariogram}
\docType{data}
\title{
Variogram of radioactivePlumes
}
\description{
A \code{{variogramModel}} generated from \code{kind = "finaldose"} of \code{\link{radioactivePlumes}} with the code given in the examples of \code{\link{fitMedianVariogram}}.
}
\usage{data("medianVariogram")}
\author{ Kristina B. Helle, \email{kristina.helle@uni-muenster.de} }
\examples{
data(medianVariogram)
}
\keyword{datasets}
