\name{optimalSD}

\alias{SDgenetic}
\alias{SDglobal}
\alias{SDgreedy}
\alias{SDmanual}
\alias{SDssa}

\docType{data}

\title{
Optimised sampling designs
}
\description{
For each of the optimisation algorithms a resulting sampling design is provided. These are taken from the examples of the respective cost functions.
}
\usage{
data(SDgenetic)
data(SDglobal)
data(SDgreedy)
data(SDmanual)
data(SDssa)
}
\format{
 \code{SDgenetic}: \code{list}  \cr
 \code{SDglobal}: \code{list}   \cr
 \code{SDgreedy}: \code{list}   \cr
 \code{SDmanual}: \code{list}   \cr
 \code{SDssa}: \code{list}      \cr 
}
\author{ Kristina B. Helle, \email{kristina.helle@uni-muenster.de} }
\examples{
data(SDgenetic)
}
\keyword{datasets}
