\name{radioactivePlumes_area}

\alias{radioactivePlumes_area}
\docType{data}

\title{
Simulations of radioactive plumes
}
\description{
The simulations show maps of pollution spreading from a single source. 
The data consists of 10 subsets, each for a different 200 km x 200 km region with one source within or in the neightbouhood to represent different climates and topographies of Europe. 
The simulations were generated by the atmospheric dispersion model RIMPUFF. 
They show radioactivity after a release of 6 nuclids, of about 1e+13 Bq/s each, during 12 hours. The nuclids represent typical half-lives and deposition characteristics to simplify release patterns that are used in radiological emergency simulations.
To represent a large variety of possible wind fields, weather data from a full year was used, a new plume starting every 36 hours and then simulations run for a week. For each subset the weather data around a different source location was used.
The simulations were generated within the EU FP7 project DETECT, for details see the reference.
}

\details{
The three kinds of the \code{values} are summaries over plumes in time. For each location they give: \cr
\code{finaldose}: total dose after a week, in Sv;\cr
\code{maxdose}: maximal hourly average dose within the week, in Sv/h;\cr
\code{time}: time until the hourly average dose first reaches 1e-7 Sv/h, in seconds.\cr
\code{maxdose} and \code{time} may be used to determine where and when a plume can be detected at a detection level of 1e-7 Sv/h. \code{finaldose} is of interest for total impact and may for example been combined with population data.
\cr

The \code{name} of \code{plumes} indicates the region subset and the date when it was started.
The exact locations are secret according to the rules of the DETECT project from which the data was taken. Thus the location parameters have been anonymised and values from different locations have been put together in one Simulations object. \cr
A: flat, coastal area in marine middle Europe with source in the centre \cr
B: hilly area in marine middle Europe with source in the centre \cr
C: mountaneous area with large valley in middle Europe with source in the centre \cr
D: hilly area in marine middle-northern Europe with source 700 km to the west of the centre, separated by sea \cr
E: flat area in northern Europe with source 400 km to the west, separated by sea \cr
F: hilly area in eastern Europe with a mountain ridge (west to east) with source 200 km to the south, separated by flat land \cr
G: mountainous area in middle Europe with source in NW corner \cr
H: hilly-mountaneous area in southern Europe with source in SW corner \cr
I: hilly area in middle Europe with source in SW corner \cr
J: flat area in north-eastern Europe with source in SW corner \cr
Note that for subsets D - J for some weather conditions the plumes did not reach the region, therefore there are less than 292 plumes in these subsets.
}
\usage{
data(radioactivePlumes_area)
}
\format{
 \code{radioactivePlumes_area}: \code{\link{Simulations}}           
}

\references{
K.B. Helle, T.O. M{\"u}ller, P. Astrup, J.E. Dyve (2014). Automatic Optimisation of Gamma Dose Rate Sensor Networks: The DETECT Optimisation Tool. Computers \& Geosciences (66), 158-167.
}


\examples{
data(radioactivePlumes_area)
# data are taken from files, the paths need to be updated
# radioactivePlumes_area = changeSimulationsPath(RadioactivePlumes_area,
# c(paste0(path.package("sensors4plumes"),"/extdata/radioactivePlumes_area_finaldose.grd"),
#  paste0(path.package("sensors4plumes"),"/extdata/radioactivePlumes_area_maxdose.grd"),
#  paste0(path.package("sensors4plumes"),"/extdata/radioactivePlumes_area_time.grd")))

spplot(RadioactivePlumes_area, zcol = 2)
}
\keyword{datasets}
