\name{SDoptimal}

\alias{SDgenetic}
\alias{SDglobal}
\alias{SDgreedy}
\alias{SDmanual}
\alias{SDssa}
\docType{data}

\title{
Optimal sampling designs
}
\description{
These are the results of sampling design optimisations by all provided optimisation algorithms. They have been generated by the code in the respective examples in \code{\link{optimiseSD_genetic}}, \code{\link{optimiseSD_global}}, \code{\link{optimiseSD_greedy}}, \code{\link{optimiseSD_manual}}, \code{\link{optimiseSD_ssa}}. The format depends on the optimisation algorithm.
}

\usage{
data(SDgenetic)
data(SDglobal)
data(SDgreedy)
data(SDmanual)
data(SDssa)
}
\format{
 The format depends on the optimisation algorithm.
}
\keyword{datasets}
