\name{SimulationsSmall}

\alias{SimulationsSmall}
\docType{data}

\title{
Artificial, small test data.
}
\description{
From example in Simulations.Rd (with random values).
}
\usage{
data(SimulationsSmall)
}
\format{
 \code{SimulationsSmall}: \code{\link{Simulations}}           
}
\author{ Kristina B. Helle, \email{kristina.helle@uni-muenster.de} }

\examples{
data(SimulationsSmall)
}
\keyword{datasets}
